Rework GPIO interface and tidy up tests.
GPIO interface is now more efficient memory wise, the design is more
straightforward, and there is now a clear distinction between the role
of a Pin and the role of a Gpio. Renamed 'make_gpio' and 'set_function'
to 'claim_as_gpio' and 'claim_as', for better clarity and readability.

Implemented the hil::Pin interface on Gpio, not Pin, allowing the GPIO
userspace driver to be initialized properly, and to make it clear that
those pins are being claimed.